Move |LockedQueue::Index| implementation into separate header.
